rename package from selenium to playwright
