+++ mrchecker-playwright-junit-example/src/main/java/com/capgemini/mrchecker/playwright/example/base/BasePageGUI.java
@@ -1,4 +1,4 @@
-package com.capgemini.mrchecker.selenium.example.base;
+package com.capgemini.mrchecker.playwright.example.base;
 
 
 import com.capgemini.mrchecker.playwright.core.BasePage;
