+++ mrchecker-playwright-junit-example/src/main/java/com/capgemini/mrchecker/playwright/example/page/DemoQAFormPage.java
@@ -1,0 +1,12 @@
+package com.capgemini.mrchecker.playwright.example.page;
+
+import com.capgemini.mrchecker.playwright.example.base.BasePageGUI;
+import com.capgemini.mrchecker.playwright.example.env.GetEnvironmentParam;
+
+public class DemoQAFormPage extends BasePageGUI {
+    private final String url = GetEnvironmentParam.DEMO_QA_FORM_URL.getValue();
+
+    public void startPage() {
+        loadPage(url);
+    }
+}
