--- mrchecker-playwright-junit-example/src/main/java/com/capgemini/mrchecker/selenium/example/page/DemoQAFormPage.java
@@ -1,12 +1,0 @@
-package com.capgemini.mrchecker.selenium.example.page;
-
-import com.capgemini.mrchecker.selenium.example.base.BasePageGUI;
-import com.capgemini.mrchecker.selenium.example.env.GetEnvironmentParam;
-
-public class DemoQAFormPage extends BasePageGUI {
-    private final String url = GetEnvironmentParam.DEMO_QA_FORM_URL.getValue();
-
-    public void startPage() {
-        loadPage(url);
-    }
-}
